split cache tests to their own file
